Update BaseClient Construct Method
This commit removes all of the setup for the SDK from the BaseClient

Because the BaseClient was where the SDK was authenticating it was authenticating every API
call.

Now instead of re-authenticating the construct method takes in the api token and that is then passed into the resources that are called.