Update The ApiClient Class To Add `getConfigJsonFilePath` Method
The `getConfigJsonFilePath` method is used by the `parseConfigFile` method.

The method previously lived inside of the `BaseClient` class.

The logic of the method is:
1. It will check for to see if the `json_key_file_path` configuration for the `connection_key` passed in exist.

If it does it will return the `json_key_file_path ` from the configuration file. This is then used when authenticating
with Google.

If the `json_key_file_path` is not set under the configuration then it will return an exception noting that.