Update All Resource Calls
Previously each of the Resource calls `Drive`, `Directory`, `Gmail`, `Sheets`, `LicenseManager`, and
`Calendar` calls to pass in `$this` as the parameter instead of passing in the `connection_key` and
`connection_config` parameters.

This change is necessary to keep the way we are calling the resources when utilizing the SDK while only
authenticating with Google the one time.