Update The ApiClientFake Construct Method
This commit update the `ApiClientFake` construct method to include an extra parameter called `authenticate`.

The `authenticate` parameter will determine rather or not when running a test, if the SDK will
attempt to authenticate with Google or not. The default is set to false, this allowed for not having to
rework all of the previously written test cases.

When running unit testing this should parameter should be set to true.
  - This allows for quicker testing as well as not having to provide a valid Google OAuth JSON key.